# openapi/openapi-componentcatalog-v1.0.0.yaml
openapi: 3.0.3
info:
  title: Component Catalog REST API
  version: '1.0.0'
  description: >
    The Component Catalog API allows clients to retrieve information about CatalogItems, CatalogFilters and Files entities.
    
    Catalog and File entities also exist internally, but only referenced by id's on requests and not returned on responses.
    
    **NOTES**:
    - The OpenAPI specification file is also used to [generate](https://openapi-generator.tech/) REST client(s) and a server REST API.
    - Clients and servers generated from the same OpenAPI specification version are guaranteed to be **compatible**.
  contact:
    name: EDPCore Team
    url: https://confluence.biscrum.com/pages/viewpage.action?spaceKey=EDP&title=Welcome
servers:
  - url: http://{baseurl}/v1
    variables:
      baseurl:
        default: localhost:8080
        description: Default address for a Component Catalog's backend REST API instance.
security:
  - bearerAuth: []
...
paths:
...
  /catalog-items:
    get:
      tags:
        - CatalogItems
      summary: List of all CatalogItems.
      description: >
        Returns a list of all CatalogItems for the given Catalog identified by catalogId.<br/>
        CatalogItems referenced on a Catalog that are either invalid or non-existent are **excluded** from the response.
      operationId: getCatalogItems
      parameters:
        - name: catalogId
          in: query
          description: id for the Catalog.
          required: true
          schema:
            type: string
        - name: sortByTitle
          in: query
          description: Sort the returned CatalogItems by title, either in ascending or descending order.
          required: true
          schema:
            $ref: '#/components/schemas/SortOrder'
          example: 'asc'
      responses:
        "200":
          description: A list of valid CatalogItems.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogItem'
        "400":
          description: Invalid parameters provided on the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
        "401":
          description: Invalid client token on the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
        "403":
          description: Insufficient permissions for the client to access the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
  /catalog-items/{id}:
    get:
      tags:
        - CatalogItems
      summary: Returns the CatalogItem associated to the provided id.
      description: >
        Returns the CatalogItem associated to the provided id, unless:
        <ul>
        <li> The id is not associated to any CatalogItem. </li>
        <li> Or the associated CatalogItem is invalid and can't be processed to create a response. </li>
        </ul>
      operationId: getCatalogItemById
      parameters:
        - name: id
          in: path
          description: id for the CatalogItem.
          required: true
          schema:
            type: string
            example: 'aSdFam...yCg=='
      responses:
        "200":
          description: The CatalogItem.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItem'
        "401":
          description: Invalid client token on the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
        "403":
          description: Insufficient permissions for the client to access the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
        "404":
          description: No CatalogItem associated to the provided id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
        "422":
          description: Invalid CatalogItem associated to the provided id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: >
        Authorization via bearer token.
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CatalogItem:
      properties:
        id:
          type: string
          example: 'aSdFam...yCg=='
        path:
          type: string
          example: projects/SOMEPROJECT/repos/some-repo/raw/CatalogItem.yaml?at=refs/heads/master
        title:
          type: string
          example: An item title
        shortDescription:
          type: string
          example: This is a short description for the item
        descriptionFileId:
          type: string
          example: cHJvam...0ZXIK
        imageFileId:
          type: string
          example: cHJvam...YXN0Z
        itemSrc:
          type: string
          example: 'https://bitbucket.some-company.com/projects/SOMEPROJECT/repos/some-repo/browse/CatalogItem.yaml?at=refs/heads/master'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/CatalogItemTag'
        authors:
          type: array
          items:
            type: string
          example:
            - '@SomeAuthor'
            - '@SomeOtherAuthor'
        date:
          type: string
          format: date-time
          example: '2021-07-01T00:00:00Z'
        userActions:
          type: array
          items:
            $ref: '#/components/schemas/CatalogItemUserAction'
      required:
        - id
        - title
        - shortDescription
        - description
        - image
        - authors
        - date
      example:
        id: aSdFam...yCg==
        path: projects/SOMEPROJECT/repos/some-repo/raw/CatalogItem.yaml?at=refs/heads/master
        title: An item title
        shortDescription: This is a short description for the item
        descriptionFileId: cHJvam...0ZXIK
        imageFileId: cHJvam...YXN0Z
        itemSrc: https://bitbucket.some-company.com/projects/SOMEPROJECT/repos/some-repo/browse/CatalogItem.yaml?at=refs/heads/master
        tags:
          - label: data
            options:
              - 'some-data-option'
              - 'some-other-data-option'
        authors:
          - '@SomeAuthor'
          - '@SomeOtherAuthor'
        date: "2021-07-01T00:00:00Z"
    CatalogItemUserAction:
      properties:
        id:
          type: string
          nullable: false
          example: 'CODE'
        displayName:
          type: string
          nullable: false
          example: 'View Code'
        url:
          type: string
          nullable: true
          example: 'https://quickstarter'
        triggerMessage:
          type: string
          nullable: true
          example: 'Provisioning a component'
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/CatalogItemUserActionParameter'
      example:
        id: "PROVISION"
        triggerMessage: "Provisioning a component custom message"
        parameters:
          - name: "workflow"
            type: "string"
            required: true
            defaultValue: "9987"
            description: "Workflow to execute."
            visible: false
...
    CatalogItemUserActionParameter:
      properties:
        name:
          type: string
          nullable: false
          example: 'workflow'
        type:
          type: string
          nullable: false
          example: 'string'
        required:
          type: boolean
          example: 'true'
        defaultValue:
          type: string
          nullable: true
          example: '123'
        defaultValues:
          nullable: true
          type: array
          items:
            type: string
          example:
            - 'some-data-option'
            - 'some-other-data-option'
        options:
          nullable: true
          type: array
          items:
            type: string
          example:
            - 'some-data-option'
            - 'some-other-data-option'
        locations:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/CatalogItemUserActionParameterLocation'
        label:
          type: string
          nullable: false
          example: 'Workflow to execute.'
        placeholder:
          type: string
          nullable: true
          example: 'some placeholder for a workflow'
        hint:
          type: string
          nullable: true
          example: 'some hint for a workflow'
        visible:
          type: boolean
          example: 'true'
        validations:
            type: array
            nullable: true
            items:
                $ref: '#/components/schemas/CatalogItemUserActionParameterValidation'
      example:
        name: "workflow"
        type: "string"
        required: true
        defaultValue: "9987"
        description: "Workflow to execute."
        visible: false
    CatalogItemUserActionParameterValidation:
      properties:
        regex:
          type: string
          example: '/^[a-z\s]{0,255}$/i'
        errorMessage:
          type: string
          example: 'There is an error in the provided value, please check it and try again.'
    CatalogItemUserActionParameterLocation:
      properties:
        location:
          type: string
          example: 'EU'
        value:
          type: string
          example: '1234'
    CatalogItemTag:
      properties:
        label:
          type: string
          example: data
        options:
          type: array
          uniqueItems: true
          items:
            type: string
          example:
            - 'some-data-option'
            - 'some-other-data-option'
...
    SortOrder:
      type: string
      enum:
        - asc
        - desc
      example: asc
...
    RestErrorMessage:
      properties:
        message:
          type: string
      required:
        - message
...