# src/main/java/com/boehringer/componentcatalog/server/api/CatalogItemsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.boehringer.componentcatalog.server.api;

import com.boehringer.componentcatalog.server.model.CatalogItem;
import com.boehringer.componentcatalog.server.model.RestErrorMessage;
import com.boehringer.componentcatalog.server.model.SortOrder;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.10.0")
@Validated
@Controller
@Tag(name = "CatalogItems", description = "CatalogItems operations.")
public interface CatalogItemsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /catalog-items/{id} : Returns the CatalogItem associated to the provided id.
     * Returns the CatalogItem associated to the provided id, unless: &lt;ul&gt; &lt;li&gt; The id is not associated to any CatalogItem. &lt;/li&gt; &lt;li&gt; Or the associated CatalogItem is invalid and can&#39;t be processed to create a response. &lt;/li&gt; &lt;/ul&gt; 
     *
     * @param id id for the CatalogItem. (required)
     * @return The CatalogItem. (status code 200)
     *         or Invalid client token on the request. (status code 401)
     *         or Insufficient permissions for the client to access the resource. (status code 403)
     *         or No CatalogItem associated to the provided id. (status code 404)
     *         or Invalid CatalogItem associated to the provided id. (status code 422)
     *         or Server error. (status code 500)
     */
    @Operation(
        operationId = "getCatalogItemById",
        summary = "Returns the CatalogItem associated to the provided id.",
        description = "Returns the CatalogItem associated to the provided id, unless: <ul> <li> The id is not associated to any CatalogItem. </li> <li> Or the associated CatalogItem is invalid and can't be processed to create a response. </li> </ul> ",
        tags = { "CatalogItems" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The CatalogItem.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CatalogItem.class))
            }),
            @ApiResponse(responseCode = "401", description = "Invalid client token on the request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "403", description = "Insufficient permissions for the client to access the resource.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "404", description = "No CatalogItem associated to the provided id.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "422", description = "Invalid CatalogItem associated to the provided id.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "500", description = "Server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/catalog-items/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<CatalogItem> getCatalogItemById(
        @Parameter(name = "id", description = "id for the CatalogItem.", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : \"aSdFam...yCg==\", \"path\" : \"projects/SOMEPROJECT/repos/some-repo/raw/CatalogItem.yaml?at=refs/heads/master\", \"title\" : \"An item title\", \"shortDescription\" : \"This is a short description for the item\", \"descriptionFileId\" : \"cHJvam...0ZXIK\", \"imageFileId\" : \"cHJvam...YXN0Z\", \"itemSrc\" : \"https://bitbucket.some-company.com/projects/SOMEPROJECT/repos/some-repo/browse/CatalogItem.yaml?at=refs/heads/master\", \"tags\" : [ { \"label\" : \"data\", \"options\" : [ \"some-data-option\", \"some-other-data-option\" ] } ], \"authors\" : [ \"@SomeAuthor\", \"@SomeOtherAuthor\" ], \"date\" : \"2021-07-01T00:00:00Z\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /catalog-items : List of all CatalogItems.
     * Returns a list of all CatalogItems for the given Catalog identified by catalogId.&lt;br/&gt; CatalogItems referenced on a Catalog that are either invalid or non-existent are **excluded** from the response. 
     *
     * @param catalogId id for the Catalog. (required)
     * @param sortByTitle Sort the returned CatalogItems by title, either in ascending or descending order. (required)
     * @return A list of valid CatalogItems. (status code 200)
     *         or Invalid parameters provided on the request. (status code 400)
     *         or Invalid client token on the request. (status code 401)
     *         or Insufficient permissions for the client to access the resource. (status code 403)
     *         or Server error. (status code 500)
     */
    @Operation(
        operationId = "getCatalogItems",
        summary = "List of all CatalogItems.",
        description = "Returns a list of all CatalogItems for the given Catalog identified by catalogId.<br/> CatalogItems referenced on a Catalog that are either invalid or non-existent are **excluded** from the response. ",
        tags = { "CatalogItems" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of valid CatalogItems.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CatalogItem.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid parameters provided on the request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "401", description = "Invalid client token on the request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "403", description = "Insufficient permissions for the client to access the resource.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "500", description = "Server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/catalog-items",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<CatalogItem>> getCatalogItems(
        @NotNull @Parameter(name = "catalogId", description = "id for the Catalog.", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "catalogId", required = true) String catalogId,
        @NotNull @Parameter(name = "sortByTitle", description = "Sort the returned CatalogItems by title, either in ascending or descending order.", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "sortByTitle", required = true) SortOrder sortByTitle
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"id\" : \"aSdFam...yCg==\", \"path\" : \"projects/SOMEPROJECT/repos/some-repo/raw/CatalogItem.yaml?at=refs/heads/master\", \"title\" : \"An item title\", \"shortDescription\" : \"This is a short description for the item\", \"descriptionFileId\" : \"cHJvam...0ZXIK\", \"imageFileId\" : \"cHJvam...YXN0Z\", \"itemSrc\" : \"https://bitbucket.some-company.com/projects/SOMEPROJECT/repos/some-repo/browse/CatalogItem.yaml?at=refs/heads/master\", \"tags\" : [ { \"label\" : \"data\", \"options\" : [ \"some-data-option\", \"some-other-data-option\" ] } ], \"authors\" : [ \"@SomeAuthor\", \"@SomeOtherAuthor\" ], \"date\" : \"2021-07-01T00:00:00Z\" }, { \"id\" : \"aSdFam...yCg==\", \"path\" : \"projects/SOMEPROJECT/repos/some-repo/raw/CatalogItem.yaml?at=refs/heads/master\", \"title\" : \"An item title\", \"shortDescription\" : \"This is a short description for the item\", \"descriptionFileId\" : \"cHJvam...0ZXIK\", \"imageFileId\" : \"cHJvam...YXN0Z\", \"itemSrc\" : \"https://bitbucket.some-company.com/projects/SOMEPROJECT/repos/some-repo/browse/CatalogItem.yaml?at=refs/heads/master\", \"tags\" : [ { \"label\" : \"data\", \"options\" : [ \"some-data-option\", \"some-other-data-option\" ] } ], \"authors\" : [ \"@SomeAuthor\", \"@SomeOtherAuthor\" ], \"date\" : \"2021-07-01T00:00:00Z\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}